+++ dataprep-api/src/test/java/org/talend/dataprep/api/service/AppSettingsAPITest.java
@@ -370,6 +370,73 @@ public class AppSettingsAPITest extends ApiServiceTestBase {
         assertThat(httpImport.get("title"), is("Add HTTP dataset"));
     }
 
+    @Test
+    public void shouldCreateUrisSettings() throws Exception {
+        // when
+        final AppSettings settings = when().get("/api/settings/").as(AppSettings.class);
+
+        // then
+        final Map<String, String> mapUriSettings = settings.getUris();
+
+        assertThat(mapUriSettings.size(), is(14));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiAggregate"), is(true));
+        assertThat(mapUriSettings.get("apiAggregate"), is("/api/aggregate"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiDatasets"), is(true));
+        assertThat(mapUriSettings.get("apiDatasets"), is("/api/datasets"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiExport"), is(true));
+        assertThat(mapUriSettings.get("apiExport"), is("/api/export"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiFolders"), is(true));
+        assertThat(mapUriSettings.get("apiFolders"), is("/api/folders"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiMail"), is(true));
+        assertThat(mapUriSettings.get("apiMail"), is("/api/mail"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiPreparations"), is(true));
+        assertThat(mapUriSettings.get("apiPreparations"), is("/api/preparations"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiPreparationsPreview"), is(true));
+        assertThat(mapUriSettings.get("apiPreparationsPreview"), is("/api/preparations/preview"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiSearch"), is(true));
+        assertThat(mapUriSettings.get("apiSearch"), is("/api/search"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiSettings"), is(true));
+        assertThat(mapUriSettings.get("apiSettings"), is("/api/settings"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiTcomp"), is(true));
+        assertThat(mapUriSettings.get("apiTcomp"), is("/api/tcomp"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiTransform"), is(true));
+        assertThat(mapUriSettings.get("apiTransform"), is("/api/transform"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiTypes"), is(true));
+        assertThat(mapUriSettings.get("apiTypes"), is("/api/types"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiUpgradeCheck"), is(true));
+        assertThat(mapUriSettings.get("apiUpgradeCheck"), is("/api/upgrade/check"));
+
+        // then
+        assertThat(mapUriSettings.containsKey("apiVersion"), is(true));
+        assertThat(mapUriSettings.get("apiVersion"), is("/api/version"));
+    }
+
     private List<String> map(final List<Map<String, String>> list, final String property) {
         return list.stream().map(col -> col.get(property)).collect(toList());
     }
